Don't attempt to get device serial on Android Q., instead use Android ID. Also remove reflection access and that will be flagged in non SDK interfaces
