+++ sample/build.gradle
@@ -75,7 +75,6 @@ android {
 
     dependencies {
         implementation 'com.android.support:appcompat-v7:28.0.0'
-        implementation fileTree(dir: 'libs', include: '*.jar')
 
         //compile 'com.scottyab:secure-preferences-lib:0.1.1'
         //snapshot
